+++ tests/ValueObject/OrderedIdListTest.php
@@ -20,6 +20,7 @@ use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesNotContainId;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesNotContainSomeIds;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListIsNotEmpty;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListsMustBeEqual;
+use DigitalCraftsman\Ids\ValueObject\Exception\IdListsMustNotBeEqual;
 use PHPUnit\Framework\TestCase;
 
 /** @coversDefaultClass \DigitalCraftsman\Ids\ValueObject\OrderedIdList */
@@ -1614,7 +1615,7 @@ final class OrderedIdListTest extends TestCase
      *
      * @covers ::mustBeEqualTo
      */
-    public function must_not_be_equal_to(): void
+    public function must_be_equal_to(): void
     {
         // -- Assert
         $this->expectException(IdListsMustBeEqual::class);
@@ -1639,6 +1640,40 @@ final class OrderedIdListTest extends TestCase
 
         // -- Act
         $originalList->mustBeEqualTo($partialList);
+    }
+
+    /**
+     * @test
+     *
+     * @covers ::mustNotBeEqualTo
+     */
+    public function must_not_be_equal_to(): void
+    {
+        // -- Assert
+        $this->expectException(IdListsMustNotBeEqual::class);
+
+        // -- Arrange
+        $idAnton = UserId::generateRandom();
+        $idMarkus = UserId::generateRandom();
+        $idPaul = UserId::generateRandom();
+        $idTom = UserId::generateRandom();
+
+        $originalList = OrderedUserIdList::fromIds([
+            $idAnton,
+            $idMarkus,
+            $idPaul,
+            $idTom,
+        ]);
+
+        $sameListInSameOrder = OrderedUserIdList::fromIds([
+            $idPaul,
+            $idAnton,
+            $idMarkus,
+            $idTom,
+        ]);
+
+        // -- Act
+        $originalList->mustNotBeEqualTo($sameListInSameOrder);
     }
 
     // -- Count